Change BirthdayList component structure
